requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
langchain==0.3.25
langgraph==0.4.8
anthropic>=0.56.0
redis==5.3.0
langchain_community==0.3.26
python-dotenv==1.0.0
pydantic==2.7.4
pydantic-settings==2.1.0
python-jose==3.3.0
python-multipart==0.0.20
redis-om==0.3.5
pytest==7.4.3
httpx==0.26.0
